fix: prevent unnecessary /projects API call
- refactors BulkCodeContext and BulkCodes component
- Calls /projects and sets projects in BulkCodeContext only if needed
-- This code is moved inside the BulkCodes component
