+++ src/features/common/Layout/BulkCodeContext.tsx
@@ -9,19 +9,12 @@ import {
   useEffect,
   useCallback,
 } from 'react';
-import {
-  getAuthenticatedRequest,
-  getRequest,
-} from '../../../utils/apiRequests/api';
+import { getAuthenticatedRequest } from '../../../utils/apiRequests/api';
 import { BulkCodeMethods } from '../../../utils/constants/bulkCodeConstants';
 import { ErrorHandlingContext } from './ErrorHandlingContext';
 import { UserPropsContext } from './UserPropsContext';
-import { TENANT_ID } from '../../../utils/constants/environment';
-import i18next from '../../../../i18n';
 
-const { useTranslation } = i18next;
-
-interface APISingleProject {
+export interface APISingleProject {
   type: string;
   geometry: unknown;
   properties: {
@@ -98,7 +91,6 @@ const BulkCodeContext = createContext<BulkCodeContextInterface | null>(null);
 export const BulkCodeProvider: FC = ({ children }) => {
   const { handleError } = useContext(ErrorHandlingContext);
   const { token, contextLoaded } = useContext(UserPropsContext);
-  const { i18n } = useTranslation();
 
   const [bulkMethod, setBulkMethod] = useState<BulkCodeMethods | null>(null);
   const [planetCashAccount, setPlanetCashAccount] =
@@ -134,55 +126,6 @@ export const BulkCodeProvider: FC = ({ children }) => {
   useEffect(() => {
     fetchProfile();
   }, [fetchProfile]);
-
-  const fetchProjectList = useCallback(async () => {
-    if (planetCashAccount) {
-      try {
-        const fetchedProjects = await getRequest<APISingleProject[]>(
-          `/app/projects`,
-          handleError,
-          undefined,
-          {
-            _scope: 'map',
-            currency: planetCashAccount.currency,
-            tenant: TENANT_ID,
-            'filter[purpose]': 'trees',
-            locale: i18n.language,
-          }
-        );
-
-        // map fetchedProjects to desired form and setProject
-        if (
-          fetchedProjects &&
-          Array.isArray(fetchedProjects) &&
-          fetchedProjects.length > 0
-        ) {
-          setProjectList(
-            // Filter projects which allow donations, and store only required values in context
-            fetchedProjects
-              .filter((project) => project.properties.allowDonations)
-              .map((project) => {
-                return {
-                  guid: project.properties.id,
-                  slug: project.properties.slug,
-                  name: project.properties.name,
-                  unitCost: project.properties.unitCost,
-                  currency: project.properties.currency,
-                  purpose: project.properties.purpose,
-                  allowDonations: project.properties.allowDonations,
-                };
-              })
-          );
-        }
-      } catch (err) {
-        console.log(err);
-      }
-    }
-  }, [planetCashAccount]);
-
-  useEffect(() => {
-    fetchProjectList();
-  }, [fetchProjectList]);
 
   const value: BulkCodeContextInterface | null = useMemo(
     () => ({
